refactor: decompose tui/ui.rs into focused rendering modules (#434)
Break the 1065-line ui.rs into 5 focused modules:
- ui.rs (151 LOC): orchestration + render_input
- render_status_bar.rs (153 LOC): status bar rendering
- render_messages.rs (409 LOC): message panel with scroll/click regions
- render_popups.rs (304 LOC): autocomplete, memory modal, permissions modal
- render_debug.rs (118 LOC): debug panel with scrollable log

No behavioral changes - pure extraction refactor.
