+++ crates/cli/src/tui/mod.rs
@@ -31,6 +31,10 @@ mod layout;
 mod message;
 mod message_formatter;
 mod modal_state;
+mod render_debug;
+mod render_messages;
+mod render_popups;
+mod render_status_bar;
 mod soft_wrap;
 mod streaming_state;
 mod thinking_indicator;
